Moved DetailsModal files to "DetailsModal" path
Moved css and tsx files to "DetailsModal" path
