+++ fetec-gda/src/components/DetailsModal/DetailsModal.css
